+++ projects/packages/stats-admin/tests/php/test-stats-notices.php
@@ -39,13 +39,8 @@ class Test_Notices extends Stats_Test_Case {
 
 		Stats_Options::set_option( 'views', 2 );
 		$this->assertTrue( self::$notices->get_notices_to_show()['opt_out_new_stats'] );
-	}
 
-	/**
-	 * Test opt out new stats notice dismissed.
-	 */
-	public function test_opt_out_new_stats_notice_dismissed() {
-		self::$notices->update_notice( 'opt_out_new_stats', 'dismissed' );
+		Stats_Options::set_option( 'views', 3 );
 		$this->assertFalse( self::$notices->get_notices_to_show()['opt_out_new_stats'] );
 	}
 
@@ -58,45 +53,6 @@ class Test_Notices extends Stats_Test_Case {
 	}
 
 	/**
-	 * Test new stats feedback notice dismissed.
-	 */
-	public function test_new_stats_feedback_notice_dismissed() {
-		self::$notices->update_notice( 'new_stats_feedback', 'dismissed' );
-		self::$notices->update_notice( 'opt_out_new_stats', 'dismissed' );
-		$this->assertFalse( self::$notices->get_notices_to_show()['new_stats_feedback'] );
-		$this->assertFalse( self::$notices->get_notices_to_show()['opt_out_new_stats'] );
-	}
-
-	/**
-	 * Test new stats feedback notice dismissed.
-	 */
-	public function test_new_stats_feedback_notice_postponed() {
-		self::$notices->update_notice( 'new_stats_feedback', 'postponed', 10000 );
-		self::$notices->update_notice( 'opt_out_new_stats', 'dismissed' );
-		$stored_notices = Stats_Options::get_option( 'notices' );
-
-		$this->assertGreaterThanOrEqual( time(), $stored_notices['new_stats_feedback']['next_show_at'] );
-		$this->assertFalse( self::$notices->get_notices_to_show()['new_stats_feedback'] );
-		$this->assertFalse( self::$notices->get_notices_to_show()['opt_out_new_stats'] );
-	}
-
-	/**
-	 * Test new stats feedback notice postponed and show again.
-	 */
-	public function test_new_stats_feedback_notice_postponed_show_again() {
-		self::$notices->update_notice( 'new_stats_feedback', 'postponed', 10000 );
-		self::$notices->update_notice( 'opt_out_new_stats', 'dismissed' );
-		Stats_Options::set_option( 'views', 3 );
-
-		$stored_notices                                       = Stats_Options::get_option( 'notices' );
-		$stored_notices['new_stats_feedback']['next_show_at'] = time() - 1;
-		Stats_Options::set_option( 'notices', $stored_notices );
-
-		$this->assertTrue( self::$notices->get_notices_to_show()['new_stats_feedback'] );
-		$this->assertFalse( self::$notices->get_notices_to_show()['opt_out_new_stats'] );
-	}
-
-	/**
 	 * Test opt in new stats notice show.
 	 */
 	public function test_opt_in_new_stats_notice_show() {
@@ -106,40 +62,9 @@ class Test_Notices extends Stats_Test_Case {
 	}
 
 	/**
-	 * Test get views.
+	 * Test opt in new stats notice show.
 	 */
-	public function test_get_new_stats_views() {
-		Stats_Options::set_option( 'views', 3 );
-		$this->assertEquals( 3, self::$notices->get_new_stats_views() );
+	public function test_traffic_page_settings_hidden() {
+		$this->assertFalse( self::$notices->get_notices_to_show()['traffic_page_settings'] );
 	}
-
-	/**
-	 * Test is notice hidden.
-	 */
-	public function test_is_notice_hidden() {
-		$this->assertFalse( self::$notices->is_notice_hidden( 'opt_in_new_stats' ) );
-		self::$notices->update_notice( 'opt_in_new_stats', 'dismissed' );
-		$this->assertTrue( self::$notices->is_notice_hidden( 'opt_in_new_stats' ) );
-		self::$notices->update_notice( 'opt_in_new_stats', 'postponed' );
-		$this->assertTrue( self::$notices->is_notice_hidden( 'opt_in_new_stats' ) );
-	}
-
-	/**
-	 * Test get hidden notices.
-	 */
-	public function test_get_hidden_notices() {
-		$this->assertEmpty( self::$notices->get_hidden_notices( 'opt_in_new_stats' ) );
-		self::$notices->update_notice( 'opt_in_new_stats', 'dismissed' );
-		self::$notices->update_notice( 'new_stats_feedback', 'postponed' );
-
-		$this->assertArrayHasKey( 'opt_in_new_stats', self::$notices->get_hidden_notices() );
-		$this->assertArrayHasKey( 'new_stats_feedback', self::$notices->get_hidden_notices() );
-
-		$stored_notices                                       = Stats_Options::get_option( 'notices' );
-		$stored_notices['new_stats_feedback']['next_show_at'] = time() - 1;
-		Stats_Options::set_option( 'notices', $stored_notices );
-
-		$this->assertArrayNotHasKey( 'new_stats_feedback', self::$notices->get_hidden_notices() );
-	}
-
 }