+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/use-suggestions-from-openai.js
@@ -2,6 +2,7 @@
  * External dependencies
  */
 import { store as blockEditorStore } from '@wordpress/block-editor';
+import { serialize } from '@wordpress/blocks';
 import { useSelect, select as selectData, useDispatch } from '@wordpress/data';
 import { useEffect, useState, useRef } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
@@ -10,9 +11,9 @@ import TurndownService from 'turndown';
 /**
  * Internal dependencies
  */
+import { DEFAULT_PROMPT_TONE } from './components/tone-dropdown-control';
 import { buildPrompt } from './create-prompt';
 import { askJetpack, askQuestion } from './get-suggestion-with-stream';
-import { DEFAULT_PROMPT_TONE } from './tone-dropdown-control';
 
 const debug = debugFactory( 'jetpack-ai-assistant' );
 
@@ -33,20 +34,12 @@ export function getPartialContentToBlock( clientId ) {
 	const editor = selectData( 'core/block-editor' );
 	const index = editor.getBlockIndex( clientId );
 	const blocks = editor.getBlocks().slice( 0, index ) ?? [];
+
 	if ( ! blocks?.length ) {
 		return '';
 	}
 
-	return turndownService.turndown(
-		blocks
-			.filter( function ( block ) {
-				return block && block.attributes && block.attributes.content;
-			} )
-			.map( function ( block ) {
-				return block.attributes.content.replaceAll( '<br/>', '\n' );
-			} )
-			.join( '\n' )
-	);
+	return turndownService.turndown( serialize( blocks ) );
 }
 
 /**
@@ -63,16 +56,7 @@ export function getContentFromBlocks() {
 		return '';
 	}
 
-	return turndownService.turndown(
-		blocks
-			.filter( function ( block ) {
-				return block && block.attributes && block.attributes.content;
-			} )
-			.map( function ( block ) {
-				return block.attributes.content.replaceAll( '<br/>', '\n' );
-			} )
-			.join( '\n' )
-	);
+	return turndownService.turndown( serialize( blocks ) );
 }
 
 const useSuggestionsFromOpenAI = ( {
@@ -84,6 +68,9 @@ const useSuggestionsFromOpenAI = ( {
 	userPrompt,
 	onSuggestionDone,
 	onUnclearPrompt,
+	onModeration,
+	refreshFeatureData,
+	requireUpgrade,
 } ) => {
 	const [ isLoadingCategories, setIsLoadingCategories ] = useState( false );
 	const [ isLoadingCompletion, setIsLoadingCompletion ] = useState( false );
@@ -204,7 +191,7 @@ const useSuggestionsFromOpenAI = ( {
 		try {
 			setIsLoadingCompletion( true );
 			setWasCompletionJustRequested( true );
-			source.current = await askQuestion( prompt, postId );
+			source.current = await askQuestion( prompt, { postId, requireUpgrade } );
 		} catch ( err ) {
 			if ( err.message ) {
 				setError( { message: err.message, code: err?.code || 'unknown', status: 'error' } );
@@ -223,13 +210,14 @@ const useSuggestionsFromOpenAI = ( {
 			setWasCompletionJustRequested( false );
 		}
 
-		source?.current.addEventListener( 'done', e => {
+		source?.current?.addEventListener( 'done', e => {
 			stopSuggestion();
 			updateBlockAttributes( clientId, { content: e.detail } );
+			refreshFeatureData();
 		} );
 
-		source?.current.addEventListener( 'error_unclear_prompt', () => {
-			source?.current.close();
+		source?.current?.addEventListener( 'error_unclear_prompt', () => {
+			source?.current?.close();
 			setIsLoadingCompletion( false );
 			setWasCompletionJustRequested( false );
 			setError( {
@@ -240,8 +228,8 @@ const useSuggestionsFromOpenAI = ( {
 			onUnclearPrompt?.();
 		} );
 
-		source?.current.addEventListener( 'error_network', () => {
-			source?.current.close();
+		source?.current?.addEventListener( 'error_network', () => {
+			source?.current?.close();
 			setIsLoadingCompletion( false );
 			setWasCompletionJustRequested( false );
 			setShowRetry( true );
@@ -252,8 +240,8 @@ const useSuggestionsFromOpenAI = ( {
 			} );
 		} );
 
-		source?.current.addEventListener( 'error_service_unavailable', () => {
-			source?.current.close();
+		source?.current?.addEventListener( 'error_service_unavailable', () => {
+			source?.current?.close();
 			setIsLoadingCompletion( false );
 			setWasCompletionJustRequested( false );
 			setShowRetry( true );
@@ -267,8 +255,8 @@ const useSuggestionsFromOpenAI = ( {
 			} );
 		} );
 
-		source?.current.addEventListener( 'error_quota_exceeded', () => {
-			source?.current.close();
+		source?.current?.addEventListener( 'error_quota_exceeded', () => {
+			source?.current?.close();
 			setIsLoadingCompletion( false );
 			setWasCompletionJustRequested( false );
 			setShowRetry( false );
@@ -279,7 +267,23 @@ const useSuggestionsFromOpenAI = ( {
 			} );
 		} );
 
-		source?.current.addEventListener( 'suggestion', e => {
+		source?.current?.addEventListener( 'error_moderation', () => {
+			source?.current?.close();
+			setIsLoadingCompletion( false );
+			setWasCompletionJustRequested( false );
+			setShowRetry( false );
+			setError( {
+				code: 'error_moderation',
+				message: __(
+					'This request has been flagged by our moderation system. Please try to rephrase it and try again.',
+					'jetpack'
+				),
+				status: 'info',
+			} );
+			onModeration?.();
+		} );
+
+		source?.current?.addEventListener( 'suggestion', e => {
 			setWasCompletionJustRequested( false );
 			debug( 'fullMessage', e.detail );
 			updateBlockAttributes( clientId, { content: e.detail } );
@@ -292,7 +296,7 @@ const useSuggestionsFromOpenAI = ( {
 			return;
 		}
 
-		source?.current.close();
+		source?.current?.close();
 		setIsLoadingCompletion( false );
 		setWasCompletionJustRequested( false );
 		onSuggestionDone?.();