+++ projects/plugins/jetpack/extensions/plugins/launchpad-save-modal/launchpad-save-modal.php
@@ -21,6 +21,14 @@ function add_launchpad_options() {
 		return;
 	}
 
+	/**
+	 * Note the `site_intent` added on this Jetpack_LaunchpadSaveModal var is currently consumed in
+	 * both ETK and WBE as a temporary solution to resolve an issue where atomic sites were DoSing
+	 * themselves due to a non-existing endpoint. Since this data already existed on the window, we
+	 * used it to avoid the unnecessary requests.
+	 * https://github.com/Automattic/wp-calypso/blob/b7ba5798fddb56484f2cdaa83e9732ad32f3ca0a/apps/editing-toolkit/editing-toolkit-plugin/dotcom-fse/lib/site-intent/use-site-intent.js#L10
+	 * https://github.com/Automattic/wp-calypso/blob/b7ba5798fddb56484f2cdaa83e9732ad32f3ca0a/apps/wpcom-block-editor/src/wpcom/features/use-site-intent.js#L9
+	 */
 	$launchpad_options = array(
 		'launchpadScreenOption'       => get_option( 'launchpad_screen' ),
 		'siteIntentOption'            => get_option( 'site_intent' ),