+++ projects/js-packages/idc/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### This is a list detailing changes for the Jetpack RNA IDC package releases.
 
+## 0.10.40 - 2023-06-06
+### Changed
+- Updated package dependencies. [#31129]
+
 ## 0.10.39 - 2023-05-02
 ### Changed
 - Updated package dependencies.