+++ projects/packages/backup/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.14.0] - 2023-06-06
+### Changed
+- Update connection module to have an RNA option that updates the design [#31201]
+- Updated package dependencies. [#31129]
+
 ## [1.13.0] - 2023-05-29
 ### Added
 - Add connection screen for secondary admins [#30862]
@@ -419,6 +424,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add API endpoints and Jetpack Backup package for managing Help…
 
+[1.14.0]: https://github.com/Automattic/jetpack-backup/compare/v1.13.0...v1.14.0
 [1.13.0]: https://github.com/Automattic/jetpack-backup/compare/v1.12.17...v1.13.0
 [1.12.17]: https://github.com/Automattic/jetpack-backup/compare/v1.12.16...v1.12.17
 [1.12.16]: https://github.com/Automattic/jetpack-backup/compare/v1.12.15...v1.12.16